signature help: skip lifetimes when non-lifetime arguments are present
